Shutdown profiler thread last
This avoids attempts from other threads to communicate with
the profiler thread, which fail because it shut down.

While the exact order of shutdown operations is probably not
terribly important, its fine to shutdown the profiler thread last
because while other threads depend on it (like core resource thread),
it depends on nothing.

Signed-off-by: Simon Wülker <simon.wuelker@arcor.de>